src/www/tests/formattest.html

<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN">
<html>
  <head>
    <title>Dragonfly.Format test</title>
    <script type="text/javascript" src="../js/lib/prototype/prototype.js"></script>
    <script type="text/javascript" src="../js/lib/MochiKit/MochiKit.js"></script>
    <script type="text/javascript" src="../js/lib/MochiKit/SimpleTest.js"></script>
    <script language="javascript">Dragonfly = { };</script>
    <script type="text/javascript" src="../js/Util.js"></script>
    <script type="text/javascript" src="../js/Format.js"></script>
    <style type="text/css">
      @import url(css/test.css);
    </style>
  </head>
    
  <body>
      <pre id="test">
	<script type="text/javascript" src="js/FormatTest.js"></script>
	<script language="javascript">
	  try {
	      Dragonfly.FormatTest ({ ok: ok, is: is });
	      ok (true, "test suite finished!");
	  } catch (e) {
	      var s = "test suite failure!\n";
	      var o = { };
	      var k = null;
	      for (k in e) {
	          if (!o[k]) {
	              s += k + ": " + e[k] + "\n";
	              o[k] = e[k];
	          }
	      }
              ok (false, s);
	  }
	</script>
      </pre>
  </body>
</html>
